requirements.txt
streamlit==1.36.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
rank-bm25==0.2.2
pdfplumber==0.10.4
pytesseract==0.3.10
Pillow>=7.1.0,<11.0.0
openpyxl==3.1.2
pandas==2.2.2
beautifulsoup4==4.12.3
lxml==5.2.2
numpy==1.26.4
scikit-learn==1.5.1
tqdm==4.66.4
transformers==4.43.3
torch>=2.2.0
accelerate==0.33.0
peft==0.11.1
datasets==2.20.0
safetensors==0.4.3
pdfminer.six==20221105
python-docx==1.1.2
tabula-py==2.9.0
PyPDF2==3.0.1
